Refactor: Split Cloudflare worker.ts into focused modules #1244
worker.ts had grown past 1000 lines. Split it (behavior-preserving) into src/env.ts (Env interface), src/oauth-utils.ts (CORS/HTML/URL helpers, validation, OAuth metadata, session-token helpers, TTL constants), and src/oauth-handlers.ts (the six OAuth flow handlers). worker.ts now holds the fetch router, handleMcp, and handleHealth (1079 -> 252 lines).

Also dedupe handleMcp's three JSON-RPC error responses into a jsonRpcError helper. Verified by an independent symbol-by-symbol behavior-preservation review; tsc --noEmit clean.
